requirements.txt
streamlit
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-google-genai
chromadb==0.4.24
python-dotenv
huggingface-hub==0.23.0
sentence-transformers==2.7.0
pypdf
pysqlite3-binary